check more effects to make sure they have all their expected properties
